Use shared CDN for vendor assets
The intent here is to reuse urls for vendor assets, so that maybe someone has visited another website that uses bootstrap, jquery, or font-awesome. If this occurs, the user will reuse their cached version of that code instead of redownloading it. All local versions of vendored code is deleted, as it is no longer needed.