Fixup relative links to be root based (#36)
Since all links are relative, the html files assume they are in the root. This leaves no room for html files to be moved to provide folders for subspaces. 11ty assumes each file is the index of a folder with that name. This change is required to fix asset paths with 11ty.