Enable nunjucks and convert root page
layout nunjucks template can be extended by pages to reduce copying
header and footer code for every page. The index page is fully
refactored to extend the shared layout. Javascript files are defered in
order to group them in the head element rather than at the bottom of the
body.