Use 11ty static site generator (#35)
* initialize npm

UNLICENSED, since it is open source and we have no intention of sharing it.

* Enable 11ty

`npm start` to develop
`npm build` to build a site in _site

* Fixup relative links to be root based

Since all links are relative, the html files assume they are in the root. This leaves no room for html files to be moved to provide folders for subspaces. 11ty assumes each file is the index of a folder with that name. This change is required to fix asset paths with 11ty.

* Remove package-lock.json from diffing

* Ignore package-lock.json in github lines changed

This makes the github line count diff more accurate of work done rather
than dependencies changed.

* Enable nunjucks and convert root page

layout nunjucks template can be extended by pages to reduce copying
header and footer code for every page. The index page is fully
refactored to extend the shared layout. Javascript files are defered in
order to group them in the head element rather than at the bottom of the
body.

* Convert /learn to nunjucks

secondary.css is a commonly used styles file, so we default it in the
template. The index nunjucks file overrides to use primary.css anyway.

* Correct Contact section padding

While converting about.html to nunjucks, it was noticed that the about
page had the correct padding for top and bottom of the Contact section,
but the main page and learn did not. The correct css class was moved to
the shared template to resolve it on all pages.

* Convert /about to nunjucks

I noticed an image darkness difference on my local, but after
downloading the live version and comparing, it might be CDN
optimization.

* Convert /volunteer to nunjucks

Spacing around Contact section is noticably smaller, but it now matches
the other pages.

* Configure netlify for 11ty

11ty builds the site if `npm build` is called. It publishes the built site for
static hosting in the `_site` folder. Currently, we know all of this
works on node 12, so we lock the version of the builder to node 12.

* Convert /sponsor to nunjucks

* Convert /assistance to nunjucks

* Convert /slack to nunjucks

* Convert /knoxville to nunjucks

City pages have different 'contact' sections, so this became an
overridable block. It is clear that multiple city pages use the same
contact form, but the form does not work, so it is not yet worth creating a
partial that can be imported as needed.

* Convert /memphis to nunjucks

* Convert /montgomery to nunjucks

* Convert /northmississippi to nunjucks

Pretty standard at this point. There's overlap with other pages, but
these are all inactive pages at the moment, so not spending time on it.

* Organize source files

Move input files to `src` directory (pretty standard)
Move output files to `dist` (pretty standard)
Move non-processed static assets to `public` (pretty standard)